requirements.txt
fastapi
pyairtable
python-dotenv
openai
spacy
beautifulsoup4
gensim
tldextract
rake-nltk